fix: add a base class for IndexedRecord implementations.
